requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
heroku==0.1.4
idna==3.2
joblib==1.0.1
numpy==1.21.0
pandas==1.3.0
psycopg2==2.9.1
python-dateutil==1.5
pytz==2021.1
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sqlparse==0.4.1
SVC==0.1
threadpoolctl==2.2.0
urllib3==1.26.6
whitenoise==5.3.0
xlrd==2.0.1
